feat: move collector pieces required by Prom fork to new Go module
See the collector/README.md for rationales and release details.
