+++ .github/dependabot.yaml
@@ -42,18 +42,30 @@ updates:
     prefix-development: chore
     include: scope
   groups:
-    # Group dep updates into one PR as single update already updates co-located deps.
-    # Skip biggest and the most complex deps: Thanos and Prometheus.
     go-deps:
       patterns:
         - "*"
+- package-ecosystem: gomod
+  directory: /collector
+  schedule:
+    interval: weekly
+  commit-message:
+    prefix: fix
+    prefix-development: chore
+    include: scope
+  groups:
+    # Group dep updates into one PR as single update already updates co-located deps.
+    # Skip biggest and the most complex deps.
+    go-deps:
+      patterns:
+      - "*"
       exclude-patterns:
-        # All of below deps has to be updated carefully as they are large, impacting
-        # our ability to import export pkg into Prometheus forks. They also often break compatibilty.
-        - "github.com/thanos-io/thanos"
-        - "github.com/prometheus/prometheus"
-        - "k8s.io/*"
-        - "sigs.k8s.io/*"
+      # All of below deps has to be updated carefully as they are large, impacting
+      # our ability to import export pkg into Prometheus forks.
+      - "github.com/thanos-io/thanos"
+      - "github.com/prometheus/prometheus"
+      - "k8s.io/*"
+      - "sigs.k8s.io/*"
 # Duplicate config for release/0.16
 - package-ecosystem: docker
   open-pull-requests-limit: 0 # Security updates only
