+++ collector/export/gce_token_source.go
