+++ collector/export/setup/setup.go
@@ -26,7 +26,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/compute/metadata"
-	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/export"
+	"github.com/GoogleCloudPlatform/prometheus-engine/collector/export"
 	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/lease"
 	kingpin "github.com/alecthomas/kingpin/v2"
 	"github.com/go-kit/log"
