+++ collector/secrets/config.go
