+++ collector/secrets/kubernetes.go
