+++ collector/secrets/kubernetes_test.go
