+++ collector/secrets/manager.go
