+++ collector/secrets/watch.go
