+++ pkg/operator/apis/monitoring/v1/pod_config.go
@@ -21,7 +21,7 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/export"
+	"github.com/GoogleCloudPlatform/prometheus-engine/collector/export"
 	"github.com/prometheus/common/config"
 	prommodel "github.com/prometheus/common/model"
 	promconfig "github.com/prometheus/prometheus/config"
